requirements.txt
Faker
numpy 
scipy
django 
drf_yasg
psycopg2
scikit-learn
djangorestframework 
djangorestframework-simplejwt
